requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
mysql-connector-python==8.3.0 